Remove most of the old code message dispatching code
