+++ maddy.conf
@@ -20,30 +20,14 @@ smtp smtp://0.0.0.0:25 {
     # Verify that hostname in EHLO/HELO resolves to the source IP. Fail if it is not.
     filter check_source_hostname
 
-    target pipeline {
-        # Deliver all mail for @example.org into sql module storage.
-        destination example.org {
-            deliver sql
-        }
-
-        # Reject remaining recipients.
-        stop 550 "Relay disabled"
-    }
+    target dummy
 }
 
 submission smtps://0.0.0.0:465 smtp://0.0.0.0:587 {
     # Use sql module for authentication.
     auth sql
 
-    target pipeline {
-        # Deliver all mail for @example.org into sql module storage.
-        destination example.org {
-            deliver sql
-        }
-
-        # Enqueue messages for outgoing delivery.
-        deliver out_queue
-    }
+    target dummy
 }
 
 queue out_queue {
